+++ explorer/app/config/model.ts
@@ -1,4 +1,4 @@
-import { HeaderProps } from "./../components/Header/Header";
+import { HeaderProps } from "../components/Header/Header";
 
 export interface SiteConfig {
   datasources: {
